ConfigData=0x33 instead of 0xde no idea
setting it to 0xde seems to be the only way short of rebooting to
fix it, but while it always returns 0x33 at the end of the function
the next insert works

hack ConfigData if it is read as 0x33.

This is sometimes comes up when inserting the module.  Clearly
something isn't completely reseting the device, because reading
musb_read_configdata even by the end of this function it is returning
0x33.