twl4030 fix power draw after charger unplug
Subtitle, only use asleep for phy_power control, continue to run the
other suspend logic.

I expect the previous commit "twl4030_phy_suspend always mark
suspended" caused the current draw as the is_charger was still set
causing the system to continue to have the charger enabled.  This fix
is better than the original anyway because the original would call to
disable the regulators for the phy causing that to issue a warning
that it was requested to disable when already disabled.

Currently is_charger is reset, and reported in the
musb_save_ctx_and_suspend which twl4030_phy_suspend calls, but when on
a charger suspend is called while charging, and then again after
charging without a resume inbetween, causing asleep to be true in the
suspend call after unplugging from the charger preventing the charger
status from being reset.  Instead only use asleep for the phy power, I
assume the phy only needs to be powered when talking to a computer USB
port, so asleep is accurate for that.